# hash_nodes Index Optimization and Performance Tests Addition