--- a/validity-prover/migrations/20250312133639_del_index_for_hash_nodes.down.sql
+++ b/validity-prover/migrations/20250312133639_del_index_for_hash_nodes.down.sql
@@ -1,0 +1,1 @@
+CREATE INDEX idx_hash_nodes_timestamp ON public.hash_nodes USING btree (timestamp_value DESC, tag);