--- a/validity-prover/migrations/20250312133639_del_index_for_hash_nodes.up.sql
+++ b/validity-prover/migrations/20250312133639_del_index_for_hash_nodes.up.sql
@@ -1,0 +1,1 @@
+DROP INDEX IF EXISTS idx_hash_nodes_timestamp;