## Change Summary
This PR includes two main changes:

* Addition of migration scripts to remove an unused index ( `idx_hash_nodes_timestamp` ) from the `hash_nodes` table
* Addition of two performance tests for the Merkle tree implementation:
  * A test to measure the performance of SELECT query patterns (for #203 )
  * A test to measure the performance of tree reset operations (for this changes)