@xbicsan 2025-03-12
## Benchmark Results

benchmark script: `validity-prover/src/trees/merkle_tree/mod.rs:test_speed_incremental_merkle_tree_reset`

### Before
before execution benchmark:
```
 schema_name |             table_name              |                index_name                | scans | index_size | total_size | index_ratio | tuples_read | tuples_fetched
-------------+-------------------------------------+------------------------------------------+-------+------------+------------+-------------+-------------+----------------
 public      | _sqlx_migrations                    | _sqlx_migrations_pkey                    |     0 | 16 kB      | 192 kB     |        8.33 |           0 |              0
 public      | observer_block_sync_eth_block_num   | observer_block_sync_eth_block_num_pkey   |     0 | 8192 bytes | 192 kB     |        4.17 |           0 |              0
 public      | observer_deposit_sync_eth_block_num | observer_deposit_sync_eth_block_num_pkey |     0 | 8192 bytes | 192 kB     |        4.17 |           0 |              0
 public      | full_blocks                         | full_blocks_pkey                         |     0 | 8192 bytes | 192 kB     |        4.17 |           0 |              0
 public      | deposit_leaf_events                 | deposit_leaf_events_pkey                 |     0 | 8192 bytes | 192 kB     |        4.17 |           0 |              0
 public      | validity_state                      | validity_state_pkey                      |     0 | 8192 bytes | 192 kB     |        4.17 |           0 |              0
 public      | tx_tree_roots                       | tx_tree_roots_pkey                       |     0 | 8192 bytes | 192 kB     |        4.17 |           0 |              0
 public      | validity_proofs                     | validity_proofs_pkey                     |     0 | 8192 bytes | 192 kB     |        4.17 |           0 |              0
 public      | prover_tasks                        | prover_tasks_pkey                        |     0 | 8192 bytes | 192 kB     |        4.17 |           0 |              0
 public      | hash_nodes                          | hash_nodes_pkey                          |     0 | 8192 bytes | 192 kB     |        4.17 |           0 |              0
 public      | leaves                              | leaves_pkey                              |     0 | 8192 bytes | 192 kB     |        4.17 |           0 |              0
 public      | leaves_len                          | leaves_len_pkey                          |     0 | 8192 bytes | 192 kB     |        4.17 |           0 |              0
 public      | deposit_leaf_events                 | idx_deposit_leaf_events_deposit_hash     |     0 | 8192 bytes | 192 kB     |        4.17 |           0 |              0
 public      | deposit_leaf_events                 | idx_deposit_leaf_events_block_tx         |     0 | 8192 bytes | 192 kB     |        4.17 |           0 |              0
 public      | full_blocks                         | idx_full_blocks_block_tx                 |     0 | 8192 bytes | 192 kB     |        4.17 |           0 |              0
 public      | hash_nodes                          | idx_hash_nodes_timestamp                 |     0 | 8192 bytes | 192 kB     |        4.17 |           0 |              0
 public      | hash_nodes                          | idx_hash_nodes_lookup                    |     0 | 8192 bytes | 192 kB     |        4.17 |           0 |              0
 public      | leaves_len                          | idx_leaves_len_lookup                    |     0 | 8192 bytes | 192 kB     |        4.17 |           0 |              0
 public      | indexed_leaves                      | indexed_leaves_pkey                      |     0 | 8192 bytes | 192 kB     |        4.17 |           0 |              0
 public      | indexed_leaves                      | idx_indexed_leaves_lookup                |     0 | 8192 bytes | 192 kB     |        4.17 |           0 |              0
 public      | indexed_leaves                      | idx_indexed_leaves_timestamp             |     0 | 8192 bytes | 192 kB     |        4.17 |           0 |              0
 public      | prover_tasks                        | idx_prover_tasks_assigned_status         |     0 | 8192 bytes | 192 kB     |        4.17 |           0 |              0
 public      | tx_tree_roots                       | idx_tx_tree_roots_block_number           |     0 | 8192 bytes | 192 kB     |        4.17 |           0 |              0
(23 rows)
```

execute benchmark:
```
$ cargo test -r trees::merkle_tree::tests::test_speed_incremental_merkle_tree_reset -- --nocapture

running 1 test
SqlIncrementMerkleTree.reset.loop0: 1000 leaves, 32 height, 0.018423834 seconds
SqlIncrementMerkleTree.reset.loop1: 1000 leaves, 32 height, 0.01807625 seconds
test trees::merkle_tree::tests::test_speed_incremental_merkle_tree_reset has been running for over 60 seconds
SqlIncrementMerkleTree.reset.loop2: 1000 leaves, 32 height, 0.035522917 seconds
SqlIncrementMerkleTree.reset.loop3: 1000 leaves, 32 height, 0.018081166 seconds
SqlIncrementMerkleTree.reset.loop4: 1000 leaves, 32 height, 0.019275959 seconds
test trees::merkle_tree::tests::test_speed_incremental_merkle_tree_reset ... ok
```

after execution benchmark:
```
 schema_name |             table_name              |                index_name                | scans  | index_size | total_size | index_ratio | tuples_read | tuples_fetched
-------------+-------------------------------------+------------------------------------------+--------+------------+------------+-------------+-------------+----------------
 public      | hash_nodes                          | hash_nodes_pkey                          | 165000 | 2928 kB    | 9760 kB    |       30.00 |       48556 |              0
 public      | hash_nodes                          | idx_hash_nodes_lookup                    | 160000 | 5200 kB    | 9760 kB    |       53.28 |       27651 |          24660
 public      | leaves_len                          | idx_leaves_len_lookup                    |  10008 | 104 kB     | 9760 kB    |        1.07 |       16334 |          11490
 public      | leaves                              | leaves_pkey                              |   5026 | 80 kB      | 9760 kB    |        0.82 |        5967 |           2509
 public      | leaves_len                          | leaves_len_pkey                          |   5019 | 64 kB      | 9760 kB    |        0.66 |        2903 |              9
 public      | hash_nodes                          | idx_hash_nodes_timestamp                 |     28 | 1240 kB    | 9760 kB    |       12.70 |      345326 |          99008
 public      | _sqlx_migrations                    | _sqlx_migrations_pkey                    |      0 | 16 kB      | 9760 kB    |        0.16 |           0 |              0
 public      | validity_proofs                     | validity_proofs_pkey                     |      0 | 8192 bytes | 9760 kB    |        0.08 |           0 |              0
 public      | prover_tasks                        | prover_tasks_pkey                        |      0 | 8192 bytes | 9760 kB    |        0.08 |           0 |              0
 public      | tx_tree_roots                       | tx_tree_roots_pkey                       |      0 | 8192 bytes | 9760 kB    |        0.08 |           0 |              0
 public      | validity_state                      | validity_state_pkey                      |      0 | 8192 bytes | 9760 kB    |        0.08 |           0 |              0
 public      | deposit_leaf_events                 | deposit_leaf_events_pkey                 |      0 | 8192 bytes | 9760 kB    |        0.08 |           0 |              0
 public      | deposit_leaf_events                 | idx_deposit_leaf_events_deposit_hash     |      0 | 8192 bytes | 9760 kB    |        0.08 |           0 |              0
 public      | deposit_leaf_events                 | idx_deposit_leaf_events_block_tx         |      0 | 8192 bytes | 9760 kB    |        0.08 |           0 |              0
 public      | full_blocks                         | idx_full_blocks_block_tx                 |      0 | 8192 bytes | 9760 kB    |        0.08 |           0 |              0
 public      | full_blocks                         | full_blocks_pkey                         |      0 | 8192 bytes | 9760 kB    |        0.08 |           0 |              0
 public      | observer_deposit_sync_eth_block_num | observer_deposit_sync_eth_block_num_pkey |      0 | 8192 bytes | 9760 kB    |        0.08 |           0 |              0
 public      | observer_block_sync_eth_block_num   | observer_block_sync_eth_block_num_pkey   |      0 | 8192 bytes | 9760 kB    |        0.08 |           0 |              0
 public      | indexed_leaves                      | indexed_leaves_pkey                      |      0 | 8192 bytes | 9760 kB    |        0.08 |           0 |              0
 public      | indexed_leaves                      | idx_indexed_leaves_lookup                |      0 | 8192 bytes | 9760 kB    |        0.08 |           0 |              0
 public      | indexed_leaves                      | idx_indexed_leaves_timestamp             |      0 | 8192 bytes | 9760 kB    |        0.08 |           0 |              0
 public      | prover_tasks                        | idx_prover_tasks_assigned_status         |      0 | 8192 bytes | 9760 kB    |        0.08 |           0 |              0
 public      | tx_tree_roots                       | idx_tx_tree_roots_block_number           |      0 | 8192 bytes | 9760 kB    |        0.08 |           0 |              0
(23 rows)
```

### After
before execution benchmark:
```
 schema_name |             table_name              |                index_name                | scans | index_size | total_size | index_ratio | tuples_read | tuples_fetched
-------------+-------------------------------------+------------------------------------------+-------+------------+------------+-------------+-------------+----------------
 public      | _sqlx_migrations                    | _sqlx_migrations_pkey                    |     0 | 16 kB      | 184 kB     |        8.70 |           0 |              0
 public      | observer_block_sync_eth_block_num   | observer_block_sync_eth_block_num_pkey   |     0 | 8192 bytes | 184 kB     |        4.35 |           0 |              0
 public      | observer_deposit_sync_eth_block_num | observer_deposit_sync_eth_block_num_pkey |     0 | 8192 bytes | 184 kB     |        4.35 |           0 |              0
 public      | full_blocks                         | full_blocks_pkey                         |     0 | 8192 bytes | 184 kB     |        4.35 |           0 |              0
 public      | deposit_leaf_events                 | deposit_leaf_events_pkey                 |     0 | 8192 bytes | 184 kB     |        4.35 |           0 |              0
 public      | validity_state                      | validity_state_pkey                      |     0 | 8192 bytes | 184 kB     |        4.35 |           0 |              0
 public      | tx_tree_roots                       | tx_tree_roots_pkey                       |     0 | 8192 bytes | 184 kB     |        4.35 |           0 |              0
 public      | validity_proofs                     | validity_proofs_pkey                     |     0 | 8192 bytes | 184 kB     |        4.35 |           0 |              0
 public      | prover_tasks                        | prover_tasks_pkey                        |     0 | 8192 bytes | 184 kB     |        4.35 |           0 |              0
 public      | hash_nodes                          | hash_nodes_pkey                          |     0 | 8192 bytes | 184 kB     |        4.35 |           0 |              0
 public      | leaves                              | leaves_pkey                              |     0 | 8192 bytes | 184 kB     |        4.35 |           0 |              0
 public      | leaves_len                          | leaves_len_pkey                          |     0 | 8192 bytes | 184 kB     |        4.35 |           0 |              0
 public      | deposit_leaf_events                 | idx_deposit_leaf_events_deposit_hash     |     0 | 8192 bytes | 184 kB     |        4.35 |           0 |              0
 public      | deposit_leaf_events                 | idx_deposit_leaf_events_block_tx         |     0 | 8192 bytes | 184 kB     |        4.35 |           0 |              0
 public      | full_blocks                         | idx_full_blocks_block_tx                 |     0 | 8192 bytes | 184 kB     |        4.35 |           0 |              0
 public      | hash_nodes                          | idx_hash_nodes_lookup                    |     0 | 8192 bytes | 184 kB     |        4.35 |           0 |              0
 public      | leaves_len                          | idx_leaves_len_lookup                    |     0 | 8192 bytes | 184 kB     |        4.35 |           0 |              0
 public      | indexed_leaves                      | indexed_leaves_pkey                      |     0 | 8192 bytes | 184 kB     |        4.35 |           0 |              0
 public      | indexed_leaves                      | idx_indexed_leaves_lookup                |     0 | 8192 bytes | 184 kB     |        4.35 |           0 |              0
 public      | indexed_leaves                      | idx_indexed_leaves_timestamp             |     0 | 8192 bytes | 184 kB     |        4.35 |           0 |              0
 public      | prover_tasks                        | idx_prover_tasks_assigned_status         |     0 | 8192 bytes | 184 kB     |        4.35 |           0 |              0
 public      | tx_tree_roots                       | idx_tx_tree_roots_block_number           |     0 | 8192 bytes | 184 kB     |        4.35 |           0 |              0
(22 rows)
```

execute benchmark:
```
$ cargo test -r trees::merkle_tree::tests::test_speed_incremental_merkle_tree_reset -- --nocapture

running 1 test
SqlIncrementMerkleTree.reset.loop0: 1000 leaves, 32 height, 0.018655208 seconds
SqlIncrementMerkleTree.reset.loop1: 1000 leaves, 32 height, 0.016640041 seconds
test trees::merkle_tree::tests::test_speed_incremental_merkle_tree_reset has been running for over 60 seconds
SqlIncrementMerkleTree.reset.loop2: 1000 leaves, 32 height, 0.016987042 seconds
SqlIncrementMerkleTree.reset.loop3: 1000 leaves, 32 height, 0.017307375 seconds
SqlIncrementMerkleTree.reset.loop4: 1000 leaves, 32 height, 0.018168292 seconds
test trees::merkle_tree::tests::test_speed_incremental_merkle_tree_reset ... ok
```

after execution benchmark:
```
 schema_name |             table_name              |                index_name                | scans  | index_size | total_size | index_ratio | tuples_read | tuples_fetched
-------------+-------------------------------------+------------------------------------------+--------+------------+------------+-------------+-------------+----------------
 public      | hash_nodes                          | hash_nodes_pkey                          | 165024 | 2928 kB    | 7848 kB    |       37.31 |      266388 |          74256
 public      | hash_nodes                          | idx_hash_nodes_lookup                    | 160000 | 4528 kB    | 7848 kB    |       57.70 |       27092 |          24660
 public      | leaves_len                          | idx_leaves_len_lookup                    |  10007 | 104 kB     | 7848 kB    |        1.33 |       16314 |           9990
 public      | leaves                              | leaves_pkey                              |   5024 | 80 kB      | 7848 kB    |        1.02 |        7802 |           2256
 public      | leaves_len                          | leaves_len_pkey                          |   5004 | 64 kB      | 7848 kB    |        0.82 |         574 |              3
 public      | _sqlx_migrations                    | _sqlx_migrations_pkey                    |      0 | 16 kB      | 7848 kB    |        0.20 |           0 |              0
 public      | tx_tree_roots                       | tx_tree_roots_pkey                       |      0 | 8192 bytes | 7848 kB    |        0.10 |           0 |              0
 public      | validity_proofs                     | validity_proofs_pkey                     |      0 | 8192 bytes | 7848 kB    |        0.10 |           0 |              0
 public      | prover_tasks                        | prover_tasks_pkey                        |      0 | 8192 bytes | 7848 kB    |        0.10 |           0 |              0
 public      | validity_state                      | validity_state_pkey                      |      0 | 8192 bytes | 7848 kB    |        0.10 |           0 |              0
 public      | deposit_leaf_events                 | deposit_leaf_events_pkey                 |      0 | 8192 bytes | 7848 kB    |        0.10 |           0 |              0
 public      | full_blocks                         | full_blocks_pkey                         |      0 | 8192 bytes | 7848 kB    |        0.10 |           0 |              0
 public      | deposit_leaf_events                 | idx_deposit_leaf_events_deposit_hash     |      0 | 8192 bytes | 7848 kB    |        0.10 |           0 |              0
 public      | deposit_leaf_events                 | idx_deposit_leaf_events_block_tx         |      0 | 8192 bytes | 7848 kB    |        0.10 |           0 |              0
 public      | full_blocks                         | idx_full_blocks_block_tx                 |      0 | 8192 bytes | 7848 kB    |        0.10 |           0 |              0
 public      | observer_deposit_sync_eth_block_num | observer_deposit_sync_eth_block_num_pkey |      0 | 8192 bytes | 7848 kB    |        0.10 |           0 |              0
 public      | observer_block_sync_eth_block_num   | observer_block_sync_eth_block_num_pkey   |      0 | 8192 bytes | 7848 kB    |        0.10 |           0 |              0
 public      | indexed_leaves                      | indexed_leaves_pkey                      |      0 | 8192 bytes | 7848 kB    |        0.10 |           0 |              0
 public      | indexed_leaves                      | idx_indexed_leaves_lookup                |      0 | 8192 bytes | 7848 kB    |        0.10 |           0 |              0
 public      | indexed_leaves                      | idx_indexed_leaves_timestamp             |      0 | 8192 bytes | 7848 kB    |        0.10 |           0 |              0
 public      | prover_tasks                        | idx_prover_tasks_assigned_status         |      0 | 8192 bytes | 7848 kB    |        0.10 |           0 |              0
 public      | tx_tree_roots                       | idx_tx_tree_roots_block_number           |      0 | 8192 bytes | 7848 kB    |        0.10 |           0 |              0
(22 rows)
```

## Benchmark Result Summary
* measured the processing time for the reset operation, and there was almost no difference in processing time before and after the changes. Therefore, we can conclude that there was no performance degradation due to the removal of DB indexes
* Additionally, since the reset operation itself only occurs during initial startup or when an error occurs, we believe there is little value in covering it with an index
* By removing one DB index, we have reduced the storage space required for DB indexes.